doorbell-balar-gpu-integration-r2 add shared Balar packet wire ABI header
Centralize BalarCudaCallPacket_t in balar_packet_wire.h for host and firmware.
