+++ src/sst/elements/balar/Makefile.am
@@ -13,6 +13,7 @@ comp_LTLIBRARIES = libbalar.la
 libbalar_la_SOURCES = \
 	balar_consts.h \
 	balar_packet.h \
+	balar_packet_wire.h \
 	balarMMIO.cc \
 	balarMMIO.h \
 	cuda_runtime_api.h \
